requirements.txt
langchain
openai
pyyaml
langchain_community
